+++ framework/Authentication/Auth.php
@@ -9,10 +9,7 @@ class Auth
     /** Get the id of the logged in user */
     public static function id(): ?int
     {
-        if (!array_key_exists('userId', $_SESSION)) {
-            return null;
-        }
-        return $_SESSION['userId'];
+        return Session::getValue('userId');
     }
 
     /** Check if the user is logged in */